#6: Extract cell editing tests into separate class
